@@ -10,10 +10,10 @@
   },
   "dependencies": {
     "axios": "^0.27.2",
-    "next": "12.1.6",
+    "next": "^14.0.0",
     "next-redux-wrapper": "^7.0.5",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-player": "^2.10.1",
@@ -25,6 +25,6 @@
   },
   "devDependencies": {
     "eslint": "8.15.0",
-    "eslint-config-next": "12.1.6"
+    "eslint-config-next": "^14.0.0"
   }
 }
